# backend/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wellisha Commerce API",
    "version": "0.3.0",
    "description": "Implemented commerce foundation. Cognito access tokens require the configured issuer/client_id; no legacy cookie authentication. JSON bodies are limited to 16384 bytes; compressed bodies are rejected. Catalog list returns a page envelope. Four simple non-stacking same-SKU offer types share one evaluator with owned carts and order lines. Percentage discounts round down per unit before multiplying quantity. Visible clients refetch within 15 seconds or sooner at schedule boundaries. Buy-X-get-Y quantities include free units. Cart totals are not checkout quotes. Pricing administration requires wellisha/pricing.write scope AND a server-owned catalog.pricing.write permission. Checkout remains disabled by default until provider/reservation/quote acceptance. TypeScript consumers reject unsafe integer money."
  },
  "security": [
    {
      "cognitoAccessToken": []
    }
  ],
  "paths": {
...
    "/v1/checkout/quotes": {
      "post": {
        "operationId": "createCheckoutQuote",
        "security": [
          {
            "cognitoAccess": []
          }
        ],
        "parameters": [],
        "responses": {
          "201": {
            "description": "Durably committed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutQuote"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderInput"
              }
            }
          }
        }
      }
    },
    "/v1/checkout/quotes/{id}": {
      "get": {
        "operationId": "getCheckoutQuote",
        "security": [
          {
            "cognitoAccess": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Durably committed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutQuote"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
...
    "/v1/staff/orders/{id}/packages": {
      "post": {
        "operationId": "createParcel",
        "security": [
          {
            "cognitoAccess": [
              "wellisha/packing.write"
            ]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{16,100}$"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Durably committed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParcelView"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParcelInput"
              }
            }
          }
        }
      }
    },
...
    "/v1/staff/orders/{id}/operations": {
      "get": {
        "operationId": "getStaffOrderOperations",
        "security": [
          {
            "cognitoAccess": [
              "wellisha/operations.read"
            ]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Durably committed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffOrderOperations"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v1/staff/orders": {
      "get": {
        "operationId": "listStaffOrders",
        "security": [
          {
            "cognitoAccess": [
              "wellisha/operations.read"
            ]
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Durably committed response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderView"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v1/staff/shipments/{id}/cancel": {
      "post": {
        "operationId": "cancelShipment",
        "security": [
          {
            "cognitoAccess": [
              "wellisha/shipping.cancel"
            ]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Durably committed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParcelView"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipmentCancellationInput"
              }
            }
          }
        }
      }
    },
    "/v1/admin/products/{id}/bundle": {
      "get": {
        "operationId": "getBundleConfiguration",
        "security": [
          {
            "cognitoAccess": [
              "wellisha/catalog.write"
            ]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Durably committed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundleConfiguration"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      },
      "patch": {
        "operationId": "updateBundleConfiguration",
        "security": [
          {
            "cognitoAccess": [
              "wellisha/catalog.write"
            ]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Durably committed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundleConfiguration"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BundleInput"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cognitoAccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Cognito access token from configured user pool and app client."
      }
    },
    "schemas": {
...
      "OrderLineInput": {
        "type": "object",
        "required": [
          "productId",
          "quantity"
        ],
        "properties": {
          "productId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Must contain a non-whitespace character."
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "maximum": 100
          }
        },
        "additionalProperties": false
      },
      "CreateOrderInput": {
        "type": "object",
        "required": [
          "addressId",
          "items"
        ],
        "properties": {
          "addressId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Must contain a non-whitespace character."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineInput"
            },
            "minItems": 1,
            "maxItems": 50
          }
        },
        "additionalProperties": false
      },
      "OrderView": {
        "type": "object",
        "required": [
          "id",
          "totalMinor",
          "currency",
          "paymentState",
          "fulfillmentState",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "totalMinor": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "currency": {
            "type": "string"
          },
          "paymentState": {
            "type": "string"
          },
          "fulfillmentState": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "correlationId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "correlationId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "QuoteLine": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "maximum": 100
          },
          "basePriceMinor": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "discountMinor": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "totalMinor": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "priceVersion": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "required": [
          "productId",
          "quantity",
          "basePriceMinor",
          "discountMinor",
          "totalMinor",
          "priceVersion"
        ],
        "additionalProperties": false
      },
      "CheckoutQuote": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "addressId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteLine"
            }
          },
          "subtotalMinor": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "shippingMinor": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "totalMinor": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "currency": {
            "type": "string",
            "enum": [
              "INR"
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "addressId",
          "items",
          "subtotalMinor",
          "shippingMinor",
          "totalMinor",
          "currency",
          "expiresAt"
        ],
        "additionalProperties": false
      },
...
      "RefundView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "amountMinor": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "state": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "orderId",
          "amountMinor",
          "state"
        ],
        "additionalProperties": false
      },
      "ParcelInput": {
        "type": "object",
        "properties": {
          "weightGrams": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "maximum": 30000
          },
          "lengthMm": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "maximum": 2000
          },
          "widthMm": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "maximum": 2000
          },
          "heightMm": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "maximum": 2000
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineInput"
            }
          }
        },
        "required": [
          "weightGrams",
          "lengthMm",
          "widthMm",
          "heightMm",
          "items"
        ],
        "additionalProperties": false
      },
      "ParcelView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "required": [
          "id",
          "orderId",
          "state",
          "version"
        ],
        "additionalProperties": false
      },
...
      "StaffOrderLine": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "maximum": 100
          },
          "totalMinor": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "allocatedQuantity": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "maximum": 100
          }
        },
        "required": [
          "productId",
          "name",
          "quantity",
          "totalMinor",
          "allocatedQuantity"
        ],
        "additionalProperties": false
      },
      "StaffOrderOperations": {
        "type": "object",
        "properties": {
          "order": {
            "$ref": "#/components/schemas/OrderView"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StaffOrderLine"
            }
          },
          "parcels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParcelView"
            }
          },
          "refunds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RefundView"
            }
          }
        },
        "required": [
          "order",
          "items",
          "parcels",
          "refunds"
        ],
        "additionalProperties": false
      },
      "ShipmentCancellationInput": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 240
          }
        },
        "required": [
          "version",
          "reason"
        ],
        "additionalProperties": false
      },
...
      "BundleConfiguration": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineInput"
            }
          }
        },
        "required": [
          "version",
          "items"
        ],
        "additionalProperties": false
      },
      "BundleInput": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineInput"
            }
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 240
          }
        },
        "required": [
          "version",
          "items",
          "reason"
        ],
        "additionalProperties": false
      }
    },
    "responses": {
      "InvalidRequest": {
        "description": "Malformed JSON, missing required input or invalid field.",
        "headers": {
          "X-Correlation-ID": {
            "description": "Server-generated diagnostic ID.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          "Cache-Control": {
            "description": "Responses are not cached.",
            "schema": {
              "type": "string"
            },
            "example": "no-store"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "Access denied. Body is not guaranteed.",
        "headers": {
          "X-Correlation-ID": {
            "description": "Server-generated diagnostic ID.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          "Cache-Control": {
            "description": "Responses are not cached.",
            "schema": {
              "type": "string"
            },
            "example": "no-store"
          }
        }
      },
      "NotFound": {
        "description": "Resource absent, belongs to another customer, or address version is stale.",
        "headers": {
          "X-Correlation-ID": {
            "description": "Server-generated diagnostic ID.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          "Cache-Control": {
            "description": "Responses are not cached.",
            "schema": {
              "type": "string"
            },
            "example": "no-store"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Conflict": {
        "description": "Stock/state conflict, changed idempotent payload or database constraint.",
        "headers": {
          "X-Correlation-ID": {
            "description": "Server-generated diagnostic ID.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          "Cache-Control": {
            "description": "Responses are not cached.",
            "schema": {
              "type": "string"
            },
            "example": "no-store"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "Unavailable": {
        "description": "Transient database failure.",
        "headers": {
          "X-Correlation-ID": {
            "description": "Server-generated diagnostic ID.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          "Cache-Control": {
            "description": "Responses are not cached.",
            "schema": {
              "type": "string"
            },
            "example": "no-store"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    }
  }
}